requirements.txt
requests
aiohttp
cryptography
pyjwt
python-dateutil
PyNaCl